requirements.txt
# Market Intelligence Monitoring System
# 마켓 인텔리전스 모니터링 시스템

# Configuration
python-dotenv==1.1.0
pyyaml==6.0.2

# HTTP & Scraping
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.13.3

# AI Analysis
anthropic==0.84.0

# Dashboard
streamlit==1.42.0

# Utilities
python-dateutil==2.9.0
